feature: refactor app
