+++ src/modules/buyflow/components/SummaryStep/SummaryStep.css
@@ -1,0 +1,10 @@
+table {
+    margin: 0 auto;
+    border-collapse: collapse;
+}
+td,
+th {
+    padding: 10px;
+    border: 2px solid dimgray;
+    text-align: left;
+}
