+++ src/modules/buyflow/components/SummaryStep/SummaryStep.tsx
@@ -2,18 +2,28 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { useSummaryStep } from './hooks/useSummaryStep';
 import { NavigationButtons } from '../NavigationButtons';
-
+import './SummaryStep.css';
 export const SummaryStep: React.FC = () => {
   const { summaryValues, purchaseRoute } = useSummaryStep();
   return (
     <>
-      <dl>
-        {summaryValues.map((field) => (
-          <li key={field.key}>
-            {field.label}:{field.value}
-          </li>
-        ))}
-      </dl>
+      <table>
+        <thead>
+          <tr>
+            <th>Field</th>
+            <th>Value</th>
+          </tr>
+        </thead>
+        <tbody>
+          {summaryValues.map((field) => (
+            <tr key={field.key}>
+              <td>{field.label}</td>
+              <td>{field.value}</td>
+            </tr>
+          ))}
+        </tbody>
+      </table>
+
       <div>
         <Link to={purchaseRoute}>Purchase</Link>
       </div>
